+++ app/src/main/java/com/taiko/noblenote/Dialogs.kt
@@ -10,10 +10,10 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.EditText
 import android.widget.FrameLayout
+import com.taiko.noblenote.document.SFile
 import rx.Observable
 import rx.android.schedulers.AndroidSchedulers
 import rx.schedulers.Schedulers
-import java.io.File
 import java.io.IOException
 
 /**
