+++ app/src/main/java/com/taiko/noblenote/document/DocumentsContractApi19.java
@@ -1,4 +1,4 @@
-package com.taiko.noblenote.Document;
+package com.taiko.noblenote.document;
 
 import android.content.ContentResolver;
 import android.content.Context;
