+++ app/src/main/java/com/taiko/noblenote/document/IDocumentFile.kt
@@ -1,6 +1,5 @@
-package com.taiko.noblenote.Document
+package com.taiko.noblenote.document
 
-import android.content.Context
 import android.net.Uri
 
 interface IDocumentFile {
