+++ app/src/main/java/com/taiko/noblenote/document/TreeUriUtil.java
@@ -1,4 +1,4 @@
-package com.taiko.noblenote.Document;
+package com.taiko.noblenote.document;
 
 import android.content.Context;
 import android.net.Uri;
