+++ app/src/main/java/com/taiko/noblenote/editor/CABEditText.java
@@ -1,4 +1,4 @@
-package com.taiko.noblenote;
+package com.taiko.noblenote.editor;
 
 import android.content.Context;
 import android.util.AttributeSet;
