+++ app/src/main/java/com/taiko/noblenote/editor/Html.java
@@ -1,4 +1,4 @@
-package com.taiko.noblenote;
+package com.taiko.noblenote.editor;
 
 /*
  * Copyright (C) 2007 The Android Open Source Project
