+++ app/src/main/java/com/taiko/noblenote/editor/TextViewUndoRedo.java
@@ -1,4 +1,4 @@
-package com.taiko.noblenote;
+package com.taiko.noblenote.editor;
 
 import android.text.Editable;
 import android.text.Selection;
