+++ app/src/main/java/com/taiko/noblenote/extensions/TextExtensions.kt
@@ -1,4 +1,4 @@
-package com.taiko.noblenote
+package com.taiko.noblenote.extensions
 
 
 /**
